+++ src/main/java/racingcar/Application.java
@@ -1,14 +1,14 @@
 package racingcar;
 
-import racingcar.controller.domain.GameController;
-import racingcar.controller.infrastructure.RaceController;
+import racingcar.controller.domain.RaceController;
+import racingcar.controller.infrastructure.CarRaceController;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        GameController controller = new RaceController(new InputView(), new OutputView());
+        RaceController controller = new CarRaceController(new InputView(), new OutputView());
         controller.run();
     }
 }
