+++ src/main/java/racingcar/controller/domain/RaceController.java
@@ -2,7 +2,7 @@ package racingcar.controller.domain;
 
 import java.util.List;
 
-public interface GameController {
+public interface RaceController {
     void run();
 
     List<String> process(String input);
